A server and client script to publish iTunes to a webpage.
This is a client script that reads what is currently playing in your
iTunes library and pushes it to your server. The server-side script
then parses the sent data to format it for your page. This is what I
use on my WordPress blog, but you could tailor it to work with just
about anything. Read the comments in the scripts.